--- esmvaltool/diag_scripts/hydrology/pcr-globwb/config-user.yml
@@ -1,50 +1,0 @@
-###############################################################################
-# User's configuration file for the ESMValTool
-###############################################################################
----
-
-# Diagnostics create plots? [true]/false
-write_plots: false
-# Diagnositcs write NetCDF files? [true]/false
-write_netcdf: true
-# Set the console log level debug, [info], warning, error
-log_level: info
-# verbosity is deprecated and will be removed in the future
-# verbosity: 1
-# Exit on warning? true/[false]
-exit_on_warning: false
-# Plot file format? [ps]/pdf/png/eps/epsi
-output_file_type: pdf
-# Destination directory
-output_dir: ./esmvaltool_output
-# Auxiliary data directory (used for some additional datasets)
-auxiliary_data_dir: ./auxiliary_data
-# Use netCDF compression true/[false]
-compress_netcdf: false
-# Save intermediary cubes in the preprocessor true/[false]
-save_intermediary_cubes: true
-# Remove the preproc dir if all fine
-remove_preproc_dir: false
-# Run at most this many tasks in parallel null/[1]/2/3/4/..
-# Set to null to use the number of available CPUs.
-# Make sure your system has enough memory for the specified number of tasks.
-max_parallel_tasks: 1
-# Path to custom config-developer file, to customise project configurations.
-# See config-developer.yml for an example. Set to None to use the default
-config_developer_file: null
-# Get profiling information for diagnostics
-# Only available for Python diagnostics
-profile_diagnostic: false
-
-# Rootpaths to the data from different projects (lists are also possible)
-rootpath:
-  CMIP5: [~/cmip5_inputpath1, ~/cmip5_inputpath2]
-  OBS: ~/Desktop/esmvaltool/input_cmorized
-  RAWOBS: ~/Desktop/esmvaltool/input_raw
-  default: ~/Desktop/esmvaltool/test
-
-# Directory structure for input data: [default]/BADC/DKRZ/ETHZ/etc
-# See config-developer.yml for definitions.
-drs:
-  CMIP5: default
-  OBS: default
